lesson_6_mvp/requirements.worker.txt
python-telegram-bot[job-queue]==21.6
pandas==2.3.2
requests==2.32.3
pydantic==2.8.2
pydantic-settings==2.3.4
pytz==2024.1
matplotlib==3.9.0
mplfinance==0.12.10b0
numpy==1.26.4
aiohttp>=3.9
beautifulsoup4>=4.12
Pillow>=9.5
scikit-learn==1.4.2
lightgbm==4.3.0
catboost==1.2.8
ta==0.11.0
watchfiles>=0.21.0
pika>=1.3.0


